Add primaryAxisStoneSize feature
Optionally specify the length of the stone in the primary axis
direction. Useful for desktops where the app width/height can
be resized continuously and you want to vary the number of
wall divisions as the viewport gets wider (or longer).

Using this new option in such a case will be less distracting and
smoother and the primary-axis size doesn't keep increasing and
decreasing as the cross-axis size changes and number of wall
divisions changes.